Add files via upload
The University Management System is designed for any university to replace their existing manual paper based system. The project university Management system includes registration of students, storing their details into the system, and also computerized fee records as well as academic records of the students. The purpose of the project entitled as “UNIVERSITY MANAGEMENT SYSTEM” is to computerize the Front Office Management of University to develop software which is user friendly simple, fast, and cost – effective. The new system is to control the information of Students. Academic records, staff and courses availability as well as number faculty available for that particular course. I have designed the given proposed system in the JSP to automate the process of day to day activities of University like Academic record activities, Admission of New Student, Passed out students record, Assign a faculty , and finally compute the fees etc., All the control is under the administrator and the other members have the rights to just see the records not to change any transaction or entry. These services are to be provided in an efficient, cost effective manner, with the goal of reducing the time and resources currently required for such tasks.


“University Management System” is developed to computerize the following functions that are performed by the system:
1. Using this university management system user can give online registrations for the seat of the student.
2. Admin can register new student through this system.
3. Admin can also check graduate students Detail generated by the system.
4. This system is also keep academic as well as co-curricular activities record of students.
5. Admin can also check the faculty appointed to the class and vacancy of the faculty for any particular subject.
 